Move URLSessionProxy
